pyproject.toml
[tool.poetry]
name = "acephalous-assembler"
version = "0.4.0"
description = "Automated Linux ISO and disk image builder for unattended installation"
authors = ["Joseph Mienko <joseph@example.com>"]
repository = "https://github.com/josephmienko/acephalous-assembler"
readme = "README.md"
packages = [
    { include = "assembler", from = "lib/python" },
]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
argon2-cffi = "^23.1.0"
PyYAML = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
ruff = "^0.1.0"
mypy = "^1.5.0"
pylint = "^4.0.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--strict-markers --tb=short"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
select = ["E", "F", "W", "I"]
line-length = 88
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true

[tool.pylint.messages_control]
disable = [
    "wrong-import-position",
    "import-error",
    "deprecated-module",
    "invalid-name",
    "unused-import",
    "redefined-builtin",
]

[tool.pylint.format]
max-line-length = 79
